+++ src/view/components/TabList.tsx
@@ -8,7 +8,7 @@ import { memo, useContext, useState } from "react";
 import { Tab } from "../../model/Tab";
 import { isPinned, isTabGroup } from "../../model/TabContainer";
 import { WindowChild } from "../../model/Window";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../contexts/WindowsContext";
 
 import PinnedContainer from "./PinnedContainer";
 import SortableItem from "./SortableItem";
