+++ src/view/contexts/StoredWindowsContext.ts
