+++ src/view/features/popup/components/WindowTabs.tsx
@@ -5,7 +5,7 @@ import { Dispatch, useContext, useEffect } from "react";
 
 import t from "../../../../i18n/Translations";
 import { flatTabsInWindow } from "../../../../model/Window";
-import { WindowsContext } from "../../../contexts/Windows";
+import { WindowsContext } from "../../../contexts/WindowsContext";
 
 import WindowTab from "./WindowTab";
