+++ src/view/features/popup/components/WindowsContainer.tsx
@@ -2,7 +2,7 @@ import { useContext, useEffect, useState } from "react";
 
 import DragAndDropContext from "../../../components/DragAndDropContext";
 import TabList from "../../../components/TabList";
-import { WindowsContext } from "../../../contexts/Windows";
+import { WindowsContext } from "../../../contexts/WindowsContext";
 
 import WindowTabs from "./WindowTabs";
