+++ src/view/hooks/useCollapseTabGroup.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { collapseTabGroup } from "../../repository/TabGroupRepository";
 import { getWindows } from "../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../contexts/WindowsContext";
 
 export const useCollapseTabGroup = (): ((groupId: number) => Promise<void>) => {
   const { setWindows } = useContext(WindowsContext);
