+++ src/view/hooks/useMoveTab.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { moveTab } from "../../repository/TabsRepository";
 import { getWindows } from "../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../contexts/WindowsContext";
 
 export const useMoveTab = (): ((
   tabId: number,
