+++ src/view/hooks/useMoveTabGroup.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { moveTabGroup } from "../../repository/TabGroupRepository";
 import { getWindows } from "../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../contexts/WindowsContext";
 
 export const useMoveTabGroup = (): ((
   groupId: number,
