+++ src/view/hooks/useMoveTabGroupToOtherWindow.ts
@@ -3,7 +3,7 @@ import { useCallback, useContext } from "react";
 import { WindowId } from "../../model/Window";
 import { moveTabGroupToOtherWindow } from "../../repository/TabGroupRepository";
 import { getWindows } from "../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../contexts/WindowsContext";
 
 export const useMoveTabGroupToOtherWindow = (): ((
   groupId: number,
