+++ src/view/hooks/usePinTab.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { pinTab } from "../../repository/TabsRepository";
 import { getWindows } from "../../repository/WindowsRepository";
-import { WindowsContext } from "../contexts/Windows";
+import { WindowsContext } from "../contexts/WindowsContext";
 
 export const usePinTab = (): ((id: number) => Promise<void>) => {
   const { setWindows } = useContext(WindowsContext);
