+++ src/view/hooks/useSaveStoredWindow.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from "react";
 
 import { Window } from "../../model/Window";
 import { saveStoredWindow } from "../../repository/WindowsRepository";
-import { StoredWindowsContext } from "../contexts/StoredWindows";
+import { StoredWindowsContext } from "../contexts/StoredWindowsContext";
 
 export const useSaveStoredWindow = (): ((
   window: Window,
