+++ src/view/providers/ThemeProvider.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider as MaterialThemeProvider } from "@emotion/react";
 import { createTheme } from "@mui/material/styles";
 import { useContext, useMemo } from "react";
 
-import { ThemeContext } from "../contexts/Theme";
+import { ThemeContext } from "../contexts/ThemeContext";
 
 type ThemeProviderProps = {
   children: React.ReactNode;
